requirements.txt
google-generativeai
python-dotenv
Pillow
streamlit
gTTS